chore: add controller for offers create

+++ priv/repo/migrations/20211025132355_create_offers.exs
@@ -9,7 +9,7 @@ defmodule OfferHunters.Repo.Migrations.CreateOffers do
       add :promotion_link, :string
       add :image, :string
       add :expiration_date, :string
-      add :value, :decimal
+      add :value, :float
 
       add :user_id, references(:users, type: :binary_id)
